So, I've hit the limit of the amalgamation
This commit reverts MOST of the changes that I committed in updating to
SQLite 3.8.11.1.

I always knew that this strategy - modifying the amalgamation directly -
was very limited. I'm now going to modify the SQLite sources directly,
and build them directly, which should actually ease development in the
long run.

I actually made the decision to abandon the amalgamation about a month
ago, and am just getting around to it now. The `tcl.tk` domain was
actually down, which confused me quite a bit. It just so happens that
the `.tk` registrar likes redirecting domains to torrent sites.